address Ivans comments
