+++ backend/src/agents/generalist_agent.py
@@ -2,9 +2,6 @@ import logging
 from src.prompts import PromptEngine
 from src.agents import ChatAgent, chat_agent
 from src.utils import Config
-from src.utils.web_utils import (
-    answer_user_question,
-)
 import json
 
 logger = logging.getLogger(__name__)
@@ -23,9 +20,6 @@ class GeneralistAgent(ChatAgent):
         try:
             answer_to_user = await answer_user_question(utterance, self.llm, self.model)
             answer_result = json.loads(answer_to_user)
-            if answer_result["status"] == "error":
-                response = {"content": "Error in finding the answer.", "ignore_validation": "false"}
-                return json.dumps(response, indent=4)
             final_answer = json.loads(answer_result["response"]).get("answer", "")
             if not final_answer:
                 response = {"content": "Error in answer format.", "ignore_validation": "false"}
@@ -37,3 +31,25 @@ class GeneralistAgent(ChatAgent):
         except Exception as e:
             logger.error(f"Error in web_general_search_core: {e}")
             return "An error occurred while processing the search query."
+
+
+async def answer_user_question(search_query, llm, model) -> str:
+    try:
+        summariser_prompt = engine.load_prompt("generalist-answer", question=search_query)
+        response = await llm.chat(model, summariser_prompt, "")
+        return json.dumps(
+            {
+                "status": "success",
+                "response": response,
+                "error": None,
+            }
+        )
+    except Exception as e:
+        logger.error(f"Error during create search term: {e}")
+        return json.dumps(
+            {
+                "status": "error",
+                "response": None,
+                "error": str(e),
+            }
+        )
