--- backend/src/prompts/templates/answer-user-question.j2
@@ -1,4 +1,0 @@
-You are an expert in providing accurate and complete answers to user queries. Your task is to **Generate a detailed answer** to the user's question.
-
-User's question is:
-{{ question }}
