+++ backend/src/supervisors/supervisor.py
@@ -34,22 +34,23 @@ async def solve_all(intent_json) -> None:
 
 
 async def solve_task(task, scratchpad, attempt=0) -> Tuple[str, str, str]:
-    if attempt == 3:
-        agent = get_generalist_agent()
-    else:
-        agent = await get_agent_for_task(task, scratchpad)
-    if agent is None:
-        raise Exception(no_agent_response)
-    logger.info(f"Agent selected: {agent.name}")
-    logger.info(f"Task is {task}")
-    answer = await agent.invoke(task)
-    parsed_json = json.loads(answer)
-    status = parsed_json.get("status", "success")
-    ignore_validation = parsed_json.get("ignore_validation", "")
-    answer_content = parsed_json.get("content", "")
-    if (ignore_validation == "true") or await is_valid_answer(answer_content, task):
-        return (agent.name, answer_content, status)
-    return await solve_task(task, scratchpad, attempt + 1)
+    for attempt in [1, 2, 3, 4]:
+        if attempt == 4:
+            agent = get_generalist_agent()
+        else:
+            agent = await get_agent_for_task(task, scratchpad)
+        if agent is None:
+            raise Exception(no_agent_response)
+        logger.info(f"Agent selected: {agent.name}")
+        logger.info(f"Task is {task}")
+        answer = await agent.invoke(task)
+        parsed_json = json.loads(answer)
+        status = parsed_json.get("status", "success")
+        ignore_validation = parsed_json.get("ignore_validation", "")
+        answer_content = parsed_json.get("content", "")
+        if (ignore_validation == "true") or await is_valid_answer(answer_content, task):
+            return (agent.name, answer_content, status)
+    raise Exception(unsolvable_response)
 
 
 async def is_valid_answer(answer, task) -> bool:
